backend/visioncare/requirements.txt
flask
flask-cors
numpy
opencv-python
pillow
werkzeug
torch
torchvision
pandas
openpyxl
scikit-learn
